OTA-2499: Move IpUptaneSecondary and IpSecondaryDiscovery out of libaktualizr
