+++ CMakeLists.txt
@@ -59,7 +59,7 @@ unset(AKTUALIZR_CHECKED_SRCS CACHE)
 set(CMAKE_POSITION_INDEPENDENT_CODE ON)
 
 # find all required libraries
-set(BOOST_COMPONENTS log_setup log filesystem program_options)
+set(BOOST_COMPONENTS log_setup log system filesystem program_options)
 set(Boost_USE_STATIC_LIBS OFF)
 add_definitions(-DBOOST_LOG_DYN_LINK)
