+++ src/CMakeLists.txt
@@ -5,6 +5,7 @@ endif()
 add_subdirectory("libaktualizr")
 add_subdirectory("sota_tools")
 add_subdirectory("aktualizr_primary")
+add_subdirectory("libaktualizr-posix")
 add_subdirectory("aktualizr_secondary")
 add_subdirectory("aktualizr_info")
 add_subdirectory("aktualizr_repo")
