+++ src/aktualizr_check_discovery/CMakeLists.txt
@@ -2,11 +2,13 @@
 set(AKTUALIZR_CHECK_DISCOVERY_SRC main.cc )
 
 add_executable(aktualizr-check-discovery ${AKTUALIZR_CHECK_DISCOVERY_SRC})
-target_link_libraries(aktualizr-check-discovery aktualizr_static_lib ${AKTUALIZR_EXTERNAL_LIBS})
+target_link_libraries(aktualizr-check-discovery aktualizr-posix ${AKTUALIZR_EXTERNAL_LIBS})
 
 install(TARGETS aktualizr-check-discovery
         COMPONENT aktualizr
         RUNTIME DESTINATION bin)
+
+target_include_directories(aktualizr-check-discovery PUBLIC ${PROJECT_SOURCE_DIR}/src/libaktualizr-posix)
 
 aktualizr_source_file_checks(${AKTUALIZR_CHECK_DISCOVERY_SRC})
