+++ src/aktualizr_secondary/CMakeLists.txt
@@ -10,8 +10,6 @@ set(AKTUALIZR_SECONDARY_LIB_SRC
 
 add_library(aktualizr_secondary_static_lib STATIC
     ${AKTUALIZR_SECONDARY_LIB_SRC}
-    $<TARGET_OBJECTS:asn1>
-    $<TARGET_OBJECTS:asn1_lib>
     $<TARGET_OBJECTS:bootloader>
     $<TARGET_OBJECTS:crypto>
     $<TARGET_OBJECTS:jsoncpp>
@@ -21,6 +19,8 @@ add_library(aktualizr_secondary_static_lib STATIC
     $<TARGET_OBJECTS:logging>
     $<TARGET_OBJECTS:uptane>
     $<TARGET_OBJECTS:socket_activation>)
+
+target_link_libraries(aktualizr_secondary_static_lib aktualizr-posix)
 
 set(OPCUA_SECONDARY_SRC
     opcuaserver_secondary_delegate.cc
@@ -39,6 +39,7 @@ endif (BUILD_OPCUA)
 target_include_directories(aktualizr_secondary_static_lib PUBLIC
     $<TARGET_PROPERTY:socket_activation,INCLUDE_DIRECTORIES>
     $<TARGET_PROPERTY:asn1_lib,INCLUDE_DIRECTORIES>
+    ${PROJECT_SOURCE_DIR}/src/libaktualizr-posix
     )
 
 add_executable(aktualizr-secondary ${AKTUALIZR_SECONDARY_SRC})
@@ -81,6 +82,7 @@ if(BUILD_OSTREE)
 
     add_aktualizr_test(NAME aktualizr_secondary_uptane
                        SOURCES uptane_test.cc
+                       LIBRARIES aktualizr-posix
                        ARGS ${PROJECT_BINARY_DIR}/ostree_repo PROJECT_WORKING_DIRECTORY)
 else(BUILD_OSTREE)
     list(APPEND TEST_SOURCES aktualizr_secondary_discovery_test.cc uptane_test.cc)
