+++ src/libaktualizr-posix/asn1/asn1-cer.cc
@@ -1,4 +1,4 @@
-#include "asn1/asn1-cer.h"
+#include "asn1-cer.h"
 #include <algorithm>
 
 static std::string cer_encode_length(size_t len) {
