+++ src/libaktualizr-posix/asn1/messages/tlsconfig.asn1
