+++ src/libaktualizr-posix/ipsecondary_discovery_test.cc
@@ -7,9 +7,9 @@
 #include <boost/program_options.hpp>
 
 #include "config/config.h"
+#include "ipsecondarydiscovery.h"
 #include "logging/logging.h"
 #include "test_utils.h"
-#include "uptane/ipsecondarydiscovery.h"
 
 std::string port;
