+++ src/libaktualizr/CMakeLists.txt
@@ -1,5 +1,4 @@
 add_subdirectory("utilities")
-add_subdirectory("asn1")
 add_subdirectory("bootloader")
 add_subdirectory("bootstrap")
 add_subdirectory("campaign")
@@ -23,8 +22,6 @@ if (BUILD_OPCUA)
 endif (BUILD_OPCUA)
 
 add_library(aktualizr_static_lib STATIC
-    $<TARGET_OBJECTS:asn1>
-    $<TARGET_OBJECTS:asn1_lib>
     $<TARGET_OBJECTS:bootloader>
     $<TARGET_OBJECTS:bootstrap>
     $<TARGET_OBJECTS:campaign>
