+++ src/libaktualizr/uptane/secondaryfactory.cc
@@ -1,7 +1,6 @@
 #include "uptane/secondaryfactory.h"
 
 #include "logging/logging.h"
-#include "uptane/ipuptanesecondary.h"
 #include "uptane/virtualsecondary.h"
 
 #ifdef ISOTP_SECONDARY_ENABLED
@@ -21,8 +20,6 @@ std::shared_ptr<SecondaryInterface> SecondaryFactory::makeSecondary(const Second
     case SecondaryType::kLegacy:
       LOG_ERROR << "Legacy secondary support is deprecated.";
       return std::shared_ptr<SecondaryInterface>();  // NULL-equivalent
-    case SecondaryType::kIpUptane:
-      return std::make_shared<IpUptaneSecondary>(sconfig);
     case SecondaryType::kIsoTpUptane:
 #ifdef ISOTP_SECONDARY_ENABLED
       return std::make_shared<IsoTpSecondary>(sconfig);
